+++ sale_rental/data/rental_data.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  © 2014-2016 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+
+<odoo noupdate="1">
+
+<record id="route_warehouse0_rental" model="stock.location.route">
+    <field name="name">Rent</field>
+    <field name="sequence">100</field>
+    <field name="warehouse_selectable" eval="True"/>
+    <field name="product_selectable" eval="False"/>
+</record>
+
+<record id="route_warehouse0_sell_rented_product" model="stock.location.route">
+    <field name="name">Sell Rented Product</field>
+    <field name="sequence">100</field>
+</record>
+
+<record id="stock.warehouse0" model="stock.warehouse">
+    <field name="rental_allowed" eval="True"/>
+</record>
+
+</odoo>
