+++ sale_rental/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from . import product
+from . import sale_order
+from . import sale_rental
+from . import stock
