--- sale_rental/stock_view.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-    Copyright (C) 2014-2015 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
--->
-
-<openerp>
-<data>
-
-
-<record id="view_warehouse" model="ir.ui.view">
-    <field name="name">rental.stock.warehouse.form</field>
-    <field name="model">stock.warehouse</field>
-    <field name="inherit_id" ref="stock.view_warehouse" />
-    <field name="arch" type="xml">
-        <field name="wh_output_stock_loc_id" position="after">
-            <field name="rental_in_location_id"/>
-            <field name="rental_out_location_id"/>
-        </field>
-        <field name="default_resupply_wh_id" position="before">
-            <field name="rental_allowed"/>
-            <field name="rental_route_id"/>
-            <field name="sell_rented_product_route_id"/>
-        </field>
-    </field>
-</record>
-
-
-</data>
-</openerp>
