+++ sale_rental/views/sale_order.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  © 2014-2016 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+
+<odoo>
+
+
+<record id="view_order_form" model="ir.ui.view">
+    <field name="name">sale_rental.view_order_form</field>
+    <field name="model">sale.order</field>
+    <field name="inherit_id" ref="sale_start_end_dates.view_order_form" />
+    <field name="arch" type="xml">
+        <xpath expr="//field[@name='order_line']/form//field[@name='start_date']" position="before">
+            <field name="rental_type"
+                attrs="{'invisible': [('rental', '=', False)], 'required': [('rental', '=', True)]}"/>
+            <field name="extension_rental_id"
+                attrs="{'invisible': [('rental_type', '!=', 'rental_extension')], 'required': [('rental_type', '=', 'rental_extension')]}"
+                domain="[('rental_product_id', '=', product_id), ('state', 'in', ('ordered', 'out'))]" />
+            <field name="sell_rental_id"
+                attrs="{'invisible': [('can_sell_rental', '=', False)]}"
+                domain="[('rented_product_id', '=', product_id), ('state', '=', 'out')]"/>
+            <field name="rental" invisible="1"/>
+            <field name="can_sell_rental" invisible="1"/>
+        </xpath>
+        <label for="product_uom_qty" position="before">
+            <field name="rental_qty" attrs="{'invisible': [('rental', '=', False)], 'required': [('rental', '=', True)]}"/>
+        </label>
+    </field>
+</record>
+
+
+</odoo>
