+++ sale_rental/views/sale_rental.xml
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <!--
-    Copyright (C) 2014-2015 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
+  © 2014-2016 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
-<openerp>
-<data>
+<odoo>
 
 
 <record id="sale_rental_form" model="ir.ui.view">
@@ -42,7 +40,6 @@
                         <field name="product_uom" string="Unit of Measure" groups="product.group_uom"/>
                         <field name="price_subtotal" sum="Total"/>
                         <field name="state"/>
-                        <field name="invoiced"/>
                     </tree>
                 </field>
             </group>
@@ -50,7 +47,6 @@
                 <field name="out_move_id" />
                 <field name="out_state" />
                 <field name="out_picking_id" />
-<!--                <field name="prodlot_id" /> -->
             </group>
             <group name="stock-in" string="Return">
                 <field name="in_move_id" />
@@ -84,10 +80,35 @@
     </field>
 </record>
 
+<record id="sale_rental_search" model="ir.ui.view">
+    <field name="name">sale.rental.search</field>
+    <field name="model">sale.rental</field>
+    <field name="arch" type="xml">
+        <search string="Search Rentals">
+            <field name="partner_id"/>
+            <filter name="ordered" string="Ordered"
+                domain="[('state', '=', 'ordered')]"/>
+            <filter name="out" string="Out" domain="[('state', '=', 'out')]"/>
+            <filter name="sell_progress" string="Sell in progress"
+                domain="[('state', '=', 'sell_progress')]"/>
+            <filter name="sold" string="Sold" domain="[('state', '=', 'sold')]"/>
+            <filter name="in" string="Back In" domain="[('state', '=', 'in')]"/>
+            <group string="Group By" name="groupby">
+                <filter name="partner_groupby" string="Customer"
+                    context="{'group_by': 'partner_id'}"/>
+                <filter name="state_groupby" string="State"
+                    context="{'group_by': 'state'}"/>
+                <filter name="start_date_groupby" string="Start Date"
+                    context="{'group_by': 'start_date'}"/>
+            </group>
+        </search>
+    </field>
+</record>
+
+
 <record id="sale_rental_action" model="ir.actions.act_window">
     <field name="name">Rentals</field>
     <field name="res_model">sale.rental</field>
-    <field name="view_type">form</field>
     <field name="view_mode">tree,form</field>
 </record>
 
@@ -95,5 +116,4 @@
     action="sale_rental_action" sequence="90"/>
 
 
-</data>
-</openerp>
+</odoo>
