+++ sale_rental/views/stock.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  © 2014-2016 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+
+<odoo>
+
+
+<record id="view_warehouse" model="ir.ui.view">
+    <field name="name">rental.stock.warehouse.form</field>
+    <field name="model">stock.warehouse</field>
+    <field name="inherit_id" ref="stock.view_warehouse" />
+    <field name="arch" type="xml">
+        <field name="wh_output_stock_loc_id" position="after">
+            <field name="rental_view_location_id"
+                attrs="{'invisible': [('rental_allowed', '=', False)]}"/>
+            <field name="rental_in_location_id"
+                attrs="{'invisible': [('rental_allowed', '=', False)]}"/>
+            <field name="rental_out_location_id"
+                attrs="{'invisible': [('rental_allowed', '=', False)]}"/>
+        </field>
+        <field name="default_resupply_wh_id" position="before">
+            <field name="rental_allowed"/>
+            <field name="rental_route_id"
+                attrs="{'invisible': [('rental_allowed', '=', False)]}"/>
+            <field name="sell_rented_product_route_id"
+                attrs="{'invisible': [('rental_allowed', '=', False)]}"/>
+        </field>
+    </field>
+</record>
+
+
+</odoo>
