+++ sale_rental/wizard/create_rental_product_view.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <!--
-    Copyright (C) 2014-2015 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
+  © 2014-2016 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
-<openerp>
-<data>
+<odoo>
+
 
 <record id="create_rental_product_form" model="ir.ui.view">
     <field name="name">rental.product.product.form</field>
@@ -37,5 +36,5 @@
     <field name="target">new</field>
 </record>
 
-</data>
-</openerp>
+
+</odoo>
